+++ emspay/controllers/front/validation.php
@@ -1,6 +1,6 @@
 <?php
 
-require_once(_PS_MODULE_DIR_ . '/emspay/vendor/autoload.php');
+require_once(_PS_MODULE_DIR_ . '/ginger/vendor/autoload.php');
 
 class emspayValidationModuleFrontController extends ModuleFrontController
 {