+++ emspaybancontact/controllers/front/validation.php
@@ -1,9 +1,10 @@
 <?php
-require_once(_PS_MODULE_DIR_ . '/emspay/vendor/autoload.php');
 
-class emspaybancontactValidationModuleFrontController extends ModuleFrontController
+use Lib\banktwins\GingerBankValidationFrontController;
+
+require_once(_PS_MODULE_DIR_ . '/ginger/vendor/autoload.php');
+
+class emspaybancontactValidationModuleFrontController extends GingerBankValidationFrontController
 {
-    
-     use Lib\EmsPayValidationTrait;
-    
+
 }