+++ emspaygooglepay/controllers/front/payment.php
@@ -1,0 +1,10 @@
+<?php
+
+use Lib\banktwins\GingerBankPaymentFrontController;
+
+require_once(_PS_MODULE_DIR_ . '/ginger/vendor/autoload.php');
+
+class emspayGooglePayPaymentModuleFrontController extends GingerBankPaymentFrontController
+{
+
+}