+++ emspaygooglepay/emspaygooglepay.php
@@ -1,0 +1,22 @@
+<?php
+
+use Lib\banktwins\GingerBankGateway;
+use Lib\components\GingerInstallTrait;
+
+if (!defined('_PS_VERSION_')) {
+    exit;
+}
+
+require_once(_PS_MODULE_DIR_ . '/ginger/vendor/autoload.php');
+
+class emspaygooglepay extends GingerBankGateway
+{
+    use GingerInstallTrait;
+
+    public function __construct()
+    {
+        $this->name = 'emspaygooglepay';
+	    $this->method_id = 'google-pay';
+        parent::__construct();
+    }
+}