+++ emspaypaypal/controllers/front/validation.php
@@ -1,8 +1,10 @@
 <?php
 
-require_once(_PS_MODULE_DIR_ . '/emspay/vendor/autoload.php');
+use Lib\banktwins\GingerBankValidationFrontController;
 
-class emspayPaypalValidationModuleFrontController extends ModuleFrontController 
+require_once(_PS_MODULE_DIR_ . '/ginger/vendor/autoload.php');
+
+class emspayPaypalValidationModuleFrontController extends GingerBankValidationFrontController
 {
-    use Lib\EmsPayValidationTrait;
+
 }