+++ ginger/Lib/banktwins/GingerBankPaymentFrontController.php
@@ -1,0 +1,10 @@
+<?php
+namespace Lib\banktwins;
+
+
+use Lib\components\GingerPaymentFrontController;
+
+class GingerBankPaymentFrontController extends GingerPaymentFrontController
+{
+
+}